+++ common/src/main/java/nl/enjarai/doabarrelroll/DoABarrelRollClient.java
@@ -1,10 +1,7 @@
 package nl.enjarai.doabarrelroll;
 
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.util.SmoothUtil;
-import nl.enjarai.doabarrelroll.api.RollEntity;
-import nl.enjarai.doabarrelroll.api.RollMouse;
 import nl.enjarai.doabarrelroll.api.event.ClientEvents;
 import nl.enjarai.doabarrelroll.api.event.RollEvents;
 import nl.enjarai.doabarrelroll.api.event.RollGroup;
@@ -14,13 +11,9 @@ import nl.enjarai.doabarrelroll.config.ModConfig;
 import nl.enjarai.doabarrelroll.config.ModConfigServer;
 import nl.enjarai.doabarrelroll.net.register.HandshakeClientRegister;
 import nl.enjarai.doabarrelroll.flight.RotationModifiers;
-import nl.enjarai.doabarrelroll.impl.key.InputContextImpl;
 import nl.enjarai.doabarrelroll.net.HandshakeClient;
-import nl.enjarai.doabarrelroll.render.HorizonLineWidget;
-import nl.enjarai.doabarrelroll.render.MomentumCrosshairWidget;
 import nl.enjarai.doabarrelroll.util.MixinHooks;
 import nl.enjarai.doabarrelroll.util.StarFoxUtil;
-import org.joml.Vector2d;
 
 public class DoABarrelRollClient {
     public static final HandshakeClient<LimitedModConfigServer, ModConfigServer> HANDSHAKE_CLIENT = new HandshakeClient<>(
@@ -67,43 +60,7 @@ public class DoABarrelRollClient {
         StarFoxUtil.register();
     }
 
-    public static void clientTick(MinecraftClient client) {
-        InputContextImpl.getContexts().forEach(InputContextImpl::tick);
-
-        if (!isFallFlying()) {
-            clearValues();
-        }
-
-        ModKeybindings.clientTick(client);
-
-        StarFoxUtil.clientTick(client);
-    }
-
-    public static void clientDisconnect() {
-        DoABarrelRollClient.HANDSHAKE_CLIENT.reset();
-    }
-
-    public static void onRenderCrosshair(DrawContext context, float tickDelta, int scaledWidth, int scaledHeight) {
-        if (!isFallFlying()) return;
-
-        var matrices = context.getMatrices();
-        var entity = MinecraftClient.getInstance().getCameraEntity();
-        var rollEntity = ((RollEntity) entity);
-        if (entity != null) {
-            if (ModConfig.INSTANCE.getShowHorizon()) {
-                HorizonLineWidget.render(matrices, scaledWidth, scaledHeight,
-                        rollEntity.doABarrelRoll$getRoll(tickDelta), entity.getPitch(tickDelta));
-            }
-
-            if (ModConfig.INSTANCE.getMomentumBasedMouse() && ModConfig.INSTANCE.getShowMomentumWidget()) {
-                var rollMouse = (RollMouse) MinecraftClient.getInstance().mouse;
-
-                MomentumCrosshairWidget.render(matrices, scaledWidth, scaledHeight, new Vector2d(rollMouse.doABarrelRoll$getMouseTurnVec()));
-            }
-        }
-    }
-
-    private static void clearValues() {
+    public static void clearValues() {
         PITCH_SMOOTHER.clear();
         YAW_SMOOTHER.clear();
         ROLL_SMOOTHER.clear();