+++ common/src/main/java/nl/enjarai/doabarrelroll/EventCallbacksClient.java
@@ -1,0 +1,50 @@
+package nl.enjarai.doabarrelroll;
+
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.DrawContext;
+import nl.enjarai.doabarrelroll.api.RollEntity;
+import nl.enjarai.doabarrelroll.api.RollMouse;
+import nl.enjarai.doabarrelroll.config.ModConfig;
+import nl.enjarai.doabarrelroll.impl.key.InputContextImpl;
+import nl.enjarai.doabarrelroll.render.HorizonLineWidget;
+import nl.enjarai.doabarrelroll.render.MomentumCrosshairWidget;
+import nl.enjarai.doabarrelroll.util.StarFoxUtil;
+import org.joml.Vector2d;
+
+public class EventCallbacksClient {
+    public static void clientTick(MinecraftClient client) {
+        InputContextImpl.getContexts().forEach(InputContextImpl::tick);
+
+        if (!DoABarrelRollClient.isFallFlying()) {
+            DoABarrelRollClient.clearValues();
+        }
+
+        ModKeybindings.clientTick(client);
+
+        StarFoxUtil.clientTick(client);
+    }
+
+    public static void clientDisconnect() {
+        DoABarrelRollClient.HANDSHAKE_CLIENT.reset();
+    }
+
+    public static void onRenderCrosshair(DrawContext context, float tickDelta, int scaledWidth, int scaledHeight) {
+        if (!DoABarrelRollClient.isFallFlying()) return;
+
+        var matrices = context.getMatrices();
+        var entity = MinecraftClient.getInstance().getCameraEntity();
+        var rollEntity = ((RollEntity) entity);
+        if (entity != null) {
+            if (ModConfig.INSTANCE.getShowHorizon()) {
+                HorizonLineWidget.render(matrices, scaledWidth, scaledHeight,
+                        rollEntity.doABarrelRoll$getRoll(tickDelta), entity.getPitch(tickDelta));
+            }
+
+            if (ModConfig.INSTANCE.getMomentumBasedMouse() && ModConfig.INSTANCE.getShowMomentumWidget()) {
+                var rollMouse = (RollMouse) MinecraftClient.getInstance().mouse;
+
+                MomentumCrosshairWidget.render(matrices, scaledWidth, scaledHeight, new Vector2d(rollMouse.doABarrelRoll$getMouseTurnVec()));
+            }
+        }
+    }
+}