+++ common/src/main/java/nl/enjarai/doabarrelroll/mixin/client/InGameHudMixin.java
@@ -4,7 +4,7 @@ import com.llamalad7.mixinextras.sugar.Share;
 import com.llamalad7.mixinextras.sugar.ref.LocalFloatRef;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.hud.InGameHud;
-import nl.enjarai.doabarrelroll.DoABarrelRollClient;
+import nl.enjarai.doabarrelroll.EventCallbacksClient;
 import nl.enjarai.doabarrelroll.util.StarFoxUtil;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -31,7 +31,7 @@ public abstract class InGameHudMixin {
     )
     private void doABarrelRoll$renderCrosshairHead(DrawContext context, CallbackInfo ci, @Share("tickDelta") LocalFloatRef tickDeltaRef) {
         context.getMatrices().push();
-        DoABarrelRollClient.onRenderCrosshair(context, tickDeltaRef.get(), scaledWidth, scaledHeight);
+        EventCallbacksClient.onRenderCrosshair(context, tickDeltaRef.get(), scaledWidth, scaledHeight);
     }
 
     @Inject(