+++ common/src/main/java/nl/enjarai/doabarrelroll/net/register/RollSyncClient.java
@@ -1,4 +1,4 @@
-package nl.enjarai.doabarrelroll.net;
+package nl.enjarai.doabarrelroll.net.register;
 
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.util.math.MathHelper;
@@ -8,20 +8,7 @@ import nl.enjarai.doabarrelroll.api.RollEntity;
 import nl.enjarai.doabarrelroll.platform.Services;
 
 public class RollSyncClient {
-    public static void sendUpdate(RollEntity entity) {
-        if (DoABarrelRollClient.HANDSHAKE_CLIENT.hasConnected()) {
-            boolean rolling = entity.doABarrelRoll$isRolling();
-            float roll = entity.doABarrelRoll$getRoll();
-
-            var buf = DoABarrelRoll.createBuf();
-            buf.writeBoolean(rolling);
-            buf.writeFloat(roll);
-
-            Services.CLIENT_NET.sendPacket(DoABarrelRoll.ROLL_CHANNEL, buf);
-        }
-    }
-
-    public static void startListening() {
+    public static void init() {
         Services.CLIENT_NET.registerListener(DoABarrelRoll.ROLL_CHANNEL, (buf, responseSender) -> {
             var client = MinecraftClient.getInstance();
             if (client.world == null) {
@@ -41,5 +28,18 @@ public class RollSyncClient {
             rollEntity.doABarrelRoll$setRolling(isRolling);
             rollEntity.doABarrelRoll$setRoll(MathHelper.wrapDegrees(roll));
         });
+    }
+
+    public static void sendUpdate(RollEntity entity) {
+        if (DoABarrelRollClient.HANDSHAKE_CLIENT.hasConnected()) {
+            boolean rolling = entity.doABarrelRoll$isRolling();
+            float roll = entity.doABarrelRoll$getRoll();
+
+            var buf = DoABarrelRoll.createBuf();
+            buf.writeBoolean(rolling);
+            buf.writeFloat(roll);
+
+            Services.CLIENT_NET.sendPacket(DoABarrelRoll.ROLL_CHANNEL, buf);
+        }
     }
 }